tidy(ui): clean up merge visible logic
